+++ src/graph/createGraph.abstractionPath.spec.ts
@@ -1,6 +1,6 @@
 import { expect, test } from 'vitest';
 import { abstractionPath } from './createGraph';
-import { Node } from './models';
+import { Node } from '../models';
 
 test('/src/components/atoms/Button.tsx atoms = /src/components/atoms', () => {
   expect(
